Allow creation/editing of resources sans upload
Resources can now be added to collections, and edited although currently
there is no support for actual upload (and corresponding changes to
resource model - size, checksum etc).  Resources that are created are
searchable.

Resources have their own permission checks, and when creating a new
resource, the permissions from the collection are inherited.  Creating a
new resource requires write-permission on the COLLECTION, but once that
is ascertained all of the permissions are then specific to the resource.

Delete not yet implemented.